requirements.txt
langgraph>=0.2
langchain-core>=0.2
google-generativeai>=0.8
pydantic>=2
python-dotenv
